requirements.txt
# 工程预装模块
# 需要额外的python包，可直接在文件后面添加
# 请确保指定的包和版本号，可通过pip安装

blueapps-open==4.0.0rc
python-json-logger==0.1.7
requests==2.22.0
MarkupSafe==1.1.1

# django
django==3.2.4
django-celery-beat==2.2.0
django-celery-results==2.1.0
django-filter==2.4.0
django-cors-headers==3.7.0
djangorestframework==3.12.4
drf-yasg==1.20.0
drf_spectacular==0.12.0


# 版本日志
mistune==0.8.4

# jwt
pycrypto==2.6.1
PyJWT==1.7.1
cryptography==2.7

# iam
bk-iam==1.1.13

# pipeline
bamboo-pipeline==3.3.0
bamboo-engine==1.1.5

# for sentry
raven==6.1.0

# for apm
ddtrace==0.14.1

# others
mock==2.0.0
urllib3==1.25.11
pytz==2020.4
greenlet==0.4.17
jsonschema==3.2.0
botocore==1.12.253

gunicorn==19.6.0
gevent==1.2.2
celery==4.4.7
mako==1.1.3
lxml==4.6.1